issue #313 Bug fix: integer overflow in GREEDY-SEQ
SeqConfiguration.storageCost used an integer to sum up index size. This made
the size of a configuration always smaller than 2GB.

+++ extensions/sequence/src/edu/ucsc/dbtune/seq/def/SeqConfiguration.java
@@ -19,8 +19,8 @@ public class SeqConfiguration {
 		this.indices = indices;
 	}
 
-	public int storageCost() {
-		int storageCost = 0;
+	public double storageCost() {
+		double storageCost = 0;
 		for (SeqIndex index : indices)
 			storageCost += index.storageCost;
 		return storageCost;
